+++ NEWS.md
@@ -1,7 +1,10 @@
 wham 1.0.3.9000
 =========================
 
+### Minor improvements
 
+* New `save.sdrep = F` option to only save `summary(sdreport)` instead of `sdreport`. Makes saved models MUCH smaller (e.g. 2MB vs. 35MB). ([2f8875](https://github.com/timjmiller/wham/commit/2f8875323c0d6845a92444a9e7d4aaa92fe29d8d)).
+* Added `proj.opts$pct.FXSPR` option, percent of F_XSPR to use for calculating catch in projections. For example, GOM cod uses F = 75% F_XSPR, so `proj.opts$pct.FXSPR = 75`.
 
 wham 1.0.3 (2021-02-05)
 =========================